refactor: 改修のため、すべてのルートを無効化 #17
