+++ app/src/app/_disabled/(legal)/legal/[slug]/page.tsx
