+++ app/src/tests/integration/not-found.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from "@testing-library/react";
 import { describe, expect, test } from "vitest";
 
-import NotFound from "@/app/(error)/not-found";
+import NotFound from "@/app/_disabled/(error)/not-found";
 
 describe("NotFoundページのテスト", () => {
   test.skip("ページが正常に表示される", () => {
